Move VM Details' Scheduling section to new Scheduling tab
Move the "Scheduling resources and requirements" section in the VM
Details tab to the new Scheduling tab, make the section as a separate
tab.
