+++ src/utils/components/PendingChanges/utils/constants.ts
@@ -2,6 +2,7 @@ export enum VirtualMachineDetailsTab {
   Overview = '',
   Details = 'details',
   YAML = 'yaml',
+  Scheduling = 'scheduling',
   Environment = 'environment',
   Events = 'events',
   Console = 'console',
@@ -14,6 +15,7 @@ export enum VirtualMachineDetailsTabLabel {
   Overview = 'Overview',
   Details = 'Details',
   YAML = 'YAML',
+  Scheduling = 'Scheduling',
   Environment = 'Environment',
   Events = 'Events',
   Console = 'Console',
