+++ src/views/virtualmachines/details/tabs/scheduling/components/Affinity.tsx
