+++ src/views/virtualmachines/details/tabs/scheduling/components/DedicatedResources.tsx
