+++ src/views/virtualmachines/details/tabs/scheduling/components/EvictionStrategy.tsx
