+++ src/views/virtualmachines/details/tabs/scheduling/components/SchedulingSection/SchedulingSection.scss
@@ -1,0 +1,10 @@
+.vm-scheduling-section {
+  .link-icon {
+    float: left;
+    margin-left: -1.5em;
+    opacity: 0;
+    &:hover {
+      opacity: 100;
+    }
+  }
+}
